Delete TaskManager.java, added separated files for NewList and ShowLists
